Send a request to the FRED API server

Send a request to the FRED API server

fredr(endpoint, ..., to_frame = TRUE, print_req = FALSE)

Arguments

endpoint
A string representing the FRED API path of interest. See https://api.stlouisfed.org/docs/fred/.

...
A series of paramters to be used in the query. Of the form param_key = 'param_value'. Acceptable parameters are specific to the API path. See https://api.stlouisfed.org/docs/fred/.

to_frame
A boolean value indicating whether or not the response should be parsed and formatted as a tbl_df. Default is TRUE.

print_req
A boolean value indicating whether or not the request should be printed as well. Useful for debugging.

Value

A tbl_df. If to_frame is TRUE, return is a reponse object from httr::GET().

Examples

 fredr("series/observations",
       series_id = "GNPCA",
       observation_start = "1990-01-01",
       observation_end = "2000-01-01")

#> # A tibble: 11 x 4
#>    realtime_start realtime_end       date   value
#>             <chr>        <chr>      <chr>   <chr>
#>  1     2017-05-31   2017-05-31 1990-01-01  9012.6
#>  2     2017-05-31   2017-05-31 1991-01-01  8998.1
#>  3     2017-05-31   2017-05-31 1992-01-01  9314.9
#>  4     2017-05-31   2017-05-31 1993-01-01  9570.0
#>  5     2017-05-31   2017-05-31 1994-01-01  9942.9
#>  6     2017-05-31   2017-05-31 1995-01-01 10217.3
#>  7     2017-05-31   2017-05-31 1996-01-01 10606.5
#>  8     2017-05-31   2017-05-31 1997-01-01 11070.1
#>  9     2017-05-31   2017-05-31 1998-01-01 11553.1
#> 10     2017-05-31   2017-05-31 1999-01-01 12103.4
#> 11     2017-05-31   2017-05-31 2000-01-01 12608.8