Get related FRED tags given one or more tags

Get related FRED tags. Optionally, filter results by tag group, or search text. Related FRED tags are the tags assigned to series that match all tags in the tag_names parameter and no tags in the exclude_tag_names parameter.

fredr_related_tags(tag_names = NULL, exclude_tag_names = NULL,
  tag_group_id = NULL, search_text = NULL, limit = NULL, offset = NULL,
  order_by = NULL, sort_order = NULL, realtime_start = NULL,
  realtime_end = NULL)

Arguments

tag_names
A semicolon delimited string of tag names to be related to. Required parameter.

exclude_tag_names
A semicolon delimited string of tag names that series match none of. No exclusions are done by default.

tag_group_id
A string tag group id to filter tags by type. No filtering by tag group by default. Possible values are:

"freq" = Frequency

"gen" = General or Concept

"geo" = Geography

"geot" = Geography Type

"rls" = Release

"seas" = Seasonal Adjustment

"src" = Source

search_text
A string indicating the words to find matching tags with. No filtering by search words by default.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
Order results by values of the specified attribute. Possible values are:

"series_count" (default)

"popularity"

"created"

"name"

"group_id"

sort_order
A string representing the order of the resulting series, sorted by the attribute values specified by order_by. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble containing tags related to tag_names and their descriptions. Data include tag name, group ID, popularity, series count, tag creation date, and additional notes.

API Documentation

fred/related_tags

See also

Examples

fredr_related_tags(tag_names = "monetary aggregates;weekly")

#> # A tibble: 13 x 6
#>    name       group_id notes            created         popularity series_count
#>  * <chr>      <chr>    <chr>            <chr>                <int>        <int>
#>  1 nation     geot     ""               2012-02-27 10:…        100           12
#>  2 usa        geo      United States o… 2012-02-27 10:…         99           12
#>  3 frb        src      Board of Govern… 2012-02-27 10:…         79           10
#>  4 h6         rls      H.6 Money Stock… 2012-08-16 15:…         48           10
#>  5 nsa        seas     Not Seasonally … 2012-02-27 10:…        100            6
#>  6 sa         seas     Seasonally Adju… 2012-02-27 10:…         87            6
#>  7 m1         gen      M1 Money Stock   2012-02-27 10:…         46            4
#>  8 m2         gen      M2 Money Stock   2012-02-27 10:…         43            4
#>  9 currency   gen      ""               2012-02-27 10:…         63            2
#> 10 discontin… gen      ""               2012-02-27 10:…         68            2
#> 11 frb stl    src      St. Louis Fed    2012-02-27 10:…         65            2
#> 12 m3         gen      M3 Money Stock   2012-02-27 10:…         37            2
#> 13 mzm        gen      MZM Money Stock  2012-02-27 10:…         27            2

fredr_related_tags(
   tag_names = "monetary aggregates;weekly",
   tag_group_id = "gen"
 )

#> # A tibble: 6 x 6
#>   name        group_id notes         created            popularity series_count
#> * <chr>       <chr>    <chr>         <chr>                   <int>        <int>
#> 1 m1          gen      M1 Money Sto… 2012-02-27 10:18:…         46            4
#> 2 m2          gen      M2 Money Sto… 2012-02-27 10:18:…         43            4
#> 3 currency    gen      ""            2012-02-27 10:18:…         63            2
#> 4 discontinu… gen      ""            2012-02-27 10:18:…         68            2
#> 5 m3          gen      M3 Money Sto… 2012-02-27 10:18:…         37            2
#> 6 mzm         gen      MZM Money St… 2012-02-27 10:18:…         27            2