Get the series on a release of economic data

Get the series on a release of economic data

fredr_release_series(release_id = NULL, filter_variable = NULL,
  filter_value = NULL, tag_names = NULL, exclude_tag_names = NULL,
  limit = NULL, offset = NULL, order_by = NULL, sort_order = NULL,
  realtime_start = NULL, realtime_end = NULL)

Arguments

release_id
An integer ID of the release. Required parameter.

filter_variable
A string indicating which attribute to indicate the attribute that results are filtered by. Possible values include: "frequency", "units", "seasonal_adjustment". No filtering by default.

filter_value
A string giving the value of the filter_variable attribute to filter results by. filter_variable must be set. No filtering by default.

tag_names
A string indicating which series tags to match. Multiple tags can be delimited by a semicolon in a single string (e.g. `"usa;gnp"``).

exclude_tag_names
A string indicating which series tags should not be matched. Multiple tags can be delimited by a semicolon in a single string (e.g. `"usa;gnp"``).

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
A string indicating an attribute by which the results are ordered by. Possible values include:

"series_id" (default)

"title"

"units"

"frequency"

"seasonal_adjustment"

"realtime_start"

"realtime_end"

"last_updated"

"observation_start"

"observation_end"

"popularity"

"group_popularity"

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/release/series

See also

Examples

# \donttest{
fredr_release_series(release_id = 20L)

#> # A tibble: 1,000 x 16
#>    id    realtime_start realtime_end title observation_sta… observation_end
#>    <chr> <chr>          <chr>        <chr> <chr>            <chr>          
#>  1 ABS1… 2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  2 ABS15 2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  3 ABS1… 2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  4 ABS1… 2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  5 ABS5… 2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  6 ABS9… 2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  7 ABST  2019-10-24     2019-10-24   Asse… 2002-12-18       2014-11-05     
#>  8 ASBO… 2019-10-24     2019-10-24   Adju… 1986-01-08       2003-01-08     
#>  9 CPFA… 2019-10-24     2019-10-24   Comm… 2002-12-18       2010-08-25     
#> 10 CPFA… 2019-10-24     2019-10-24   Comm… 2002-12-18       2010-08-25     
#> # … with 990 more rows, and 10 more variables: frequency <chr>,
#> #   frequency_short <chr>, units <chr>, units_short <chr>,
#> #   seasonal_adjustment <chr>, seasonal_adjustment_short <chr>,
#> #   last_updated <chr>, popularity <int>, group_popularity <int>, notes <chr>

fredr_release_series(release_id = 20L, order_by = "popularity")

#> # A tibble: 1,000 x 16
#>    id    realtime_start realtime_end title observation_sta… observation_end
#>    <chr> <chr>          <chr>        <chr> <chr>            <chr>          
#>  1 H0RE… 2019-10-24     2019-10-24   Asse… 2002-12-18       2010-02-03     
#>  2 H0RE… 2019-10-24     2019-10-24   Asse… 2002-12-18       2010-01-27     
#>  3 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#>  4 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#>  5 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#>  6 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#>  7 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#>  8 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#>  9 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#> 10 RESP… 2019-10-24     2019-10-24   Asse… 2002-12-18       2019-10-16     
#> # … with 990 more rows, and 10 more variables: frequency <chr>,
#> #   frequency_short <chr>, units <chr>, units_short <chr>,
#> #   seasonal_adjustment <chr>, seasonal_adjustment_short <chr>,
#> #   last_updated <chr>, popularity <int>, group_popularity <int>, notes <chr>

# Extract the "catalog" of series from a release on a certain date
fredr_release_series(
   release_id = 20L,
   realtime_end = as.Date("2018-07-13"),
   order_by = "popularity"
)

#> # A tibble: 912 x 16
#>    id    realtime_start realtime_end title observation_sta… observation_end
#>    <chr> <chr>          <chr>        <chr> <chr>            <chr>          
#>  1 D7WA… 2011-07-07     2018-07-13   Asse… 2008-10-29       2010-08-25     
#>  2 D4WA… 2011-07-07     2018-07-13   Asse… 2008-10-29       2010-08-25     
#>  3 D10W… 2011-07-07     2018-07-13   Asse… 2008-10-29       2010-08-25     
#>  4 D8WA… 2011-07-07     2018-07-13   Asse… 2008-10-29       2010-08-25     
#>  5 D12W… 2011-07-07     2014-11-05   Asse… 2009-11-18       2014-10-29     
#>  6 D12W… 2014-11-06     2018-07-13   Asse… 2009-11-18       2014-11-05     
#>  7 D1WA… 2011-07-07     2018-07-13   Asse… 2009-12-02       2011-01-12     
#>  8 D10W… 2011-07-07     2018-07-13   Asse… 2009-12-02       2011-01-12     
#>  9 D9WA… 2011-07-07     2018-07-13   Asse… 2009-12-02       2011-01-12     
#> 10 D12W… 2011-07-07     2018-07-13   Asse… 2009-12-02       2011-01-12     
#> # … with 902 more rows, and 10 more variables: frequency <chr>,
#> #   frequency_short <chr>, units <chr>, units_short <chr>,
#> #   seasonal_adjustment <chr>, seasonal_adjustment_short <chr>,
#> #   last_updated <chr>, popularity <int>, group_popularity <int>, notes <chr>

# }