Get the sources for a release of economic data

Get the sources for a release of economic data

fredr_release_sources(
  release_id,
  ...,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

release_id

An integer ID of the release.

...

These dots only exist for future extensions and should be empty.

realtime_start

A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end

A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/release/sources

See also

Examples

if (fredr_has_key()) {
# Where does the data for ID 10 come from?
fredr_release_sources(release_id = 10L)
}
#> # A tibble: 1 × 5
#>      id realtime_start realtime_end name                            link        
#>   <int> <chr>          <chr>        <chr>                           <chr>       
#> 1    22 2023-03-13     2023-03-13   U.S. Bureau of Labor Statistics https://www…