Get release table trees for a given release

You can go directly to the tree structure by passing the appropriate element_id. You may also use a drill-down approach to start at the root (top most) element by leaving the element_id off.

fredr_release_tables(
  release_id,
  ...,
  element_id = NULL,
  include_observation_values = NULL,
  observation_date = NULL
)

Arguments

release_id
An integer ID of the release.

...
These dots only exist for future extensions and should be empty.

element_id
An integer ID for the desired release table element.

include_observation_values
A boolean indicating if observations should be returned with the release table element. Observations will only be returned for a series type element. Default is FALSE.

observation_date
A Date indicating which observation date to include with the release table. Default is 9999-12-31 (latest date available).

Value

A tibble object with nested results.

API Documentation

fred/release/tables

See also

Examples

if (fredr_has_key()) {
fredr_release_tables(release_id = 10L)

# Digging further into a release element
fredr_release_tables(release_id = 53L, element_id = 12886)
}

#> # A tibble: 4 × 2
#>   name  value           
#>   <chr> <list>          
#> 1 12887 <named list [9]>
#> 2 12888 <named list [9]>
#> 3 12889 <named list [9]>
#> 4 12890 <named list [9]>