Get all releases of economic data

Get all releases of economic data

fredr_releases(
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
Order results by values of the specified attribute. Possible values include: 'release_id' (default), 'name', 'press_release', 'realtime_start', 'realtime_end'.

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/releases

See also

Examples

# \donttest{
fredr_releases(limit = 20L)

#> # A tibble: 20 x 7
#>       id realtime_start realtime_end name      press_release link      notes    
#>    <int> <chr>          <chr>        <chr>     <lgl>         <chr>     <chr>    
#>  1     9 2020-05-21     2020-05-21   Advance … TRUE          http://w… "The U.S…
#>  2    10 2020-05-21     2020-05-21   Consumer… TRUE          http://w…  NA      
#>  3    11 2020-05-21     2020-05-21   Employme… TRUE          http://w…  NA      
#>  4    13 2020-05-21     2020-05-21   G.17 Ind… TRUE          http://w…  NA      
#>  5    14 2020-05-21     2020-05-21   G.19 Con… TRUE          http://w…  NA      
#>  6    15 2020-05-21     2020-05-21   G.5 Fore… TRUE          http://w…  NA      
#>  7    17 2020-05-21     2020-05-21   H.10 For… TRUE          http://w…  NA      
#>  8    18 2020-05-21     2020-05-21   H.15 Sel… TRUE          http://w…  NA      
#>  9    19 2020-05-21     2020-05-21   H.3 Aggr… TRUE          http://w…  NA      
#> 10    20 2020-05-21     2020-05-21   H.4.1 Fa… TRUE          http://w…  NA      
#> 11    21 2020-05-21     2020-05-21   H.6 Mone… TRUE          http://w…  NA      
#> 12    22 2020-05-21     2020-05-21   H.8 Asse… TRUE          http://w…  NA      
#> 13    25 2020-05-21     2020-05-21   Manufact… TRUE          http://w… "The Man…
#> 14    27 2020-05-21     2020-05-21   New Resi… TRUE          http://w…  NA      
#> 15    46 2020-05-21     2020-05-21   Producer… TRUE          http://w…  NA      
#> 16    47 2020-05-21     2020-05-21   Producti… TRUE          http://w…  NA      
#> 17    49 2020-05-21     2020-05-21   U.S. Int… TRUE          https://…  NA      
#> 18    50 2020-05-21     2020-05-21   Employme… TRUE          http://w…  NA      
#> 19    51 2020-05-21     2020-05-21   U.S. Int… TRUE          https://…  NA      
#> 20    52 2020-05-21     2020-05-21   Z.1 Fina… TRUE          http://w… "The Fin…

# }