Get release dates for all releases of economic data.

Get release dates for all releases of economic data. Note that release dates are published by data sources and do not necessarily represent when data will be available on the FRED or ALFRED websites.

fredr_releases_dates(
  ...,
  limit = NULL,
  offset = NULL,
  sort_order = NULL,
  order_by = NULL,
  include_release_dates_with_no_data = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

...

These dots only exist for future extensions and should be empty.

limit

An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset

An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

sort_order

A string representing the order of the resulting series. Possible values are: "asc" and "desc" (default).

order_by

Order results by values of the specified attribute. Possible values include: 'release_date' (default), 'release_id', 'release_name'.

include_release_dates_with_no_data

A boolean value indicating if the results with no data available should be returned as well. Default is FALSE.

realtime_start

A Date indicating the start of the real-time period. Defaults to the first day of the current year. For more information, see Real-Time Periods.

realtime_end

A Date indicating the end of the real-time period. Defaults to 9999-12-31 (latest available). For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/releases/dates

See also

Examples

if (fredr_has_key()) {
fredr_releases_dates(limit = 20L)
}
#> # A tibble: 20 × 3
#>    release_id release_name                                                 date 
#>         <int> <chr>                                                        <chr>
#>  1         72 Daily Treasury Inflation-Indexed Securities                  2022…
#>  2        502 Euro Short Term Rate                                         2022…
#>  3        492 SONIA Interest Rate Benchmark                                2022…
#>  4        441 Coinbase Cryptocurrencies                                    2022…
#>  5        101 FOMC Press Release                                           2022…
#>  6        441 Coinbase Cryptocurrencies                                    2022…
#>  7        101 FOMC Press Release                                           2022…
#>  8        494 Chicago Fed Advance Retail Trade Summary                     2022…
#>  9        441 Coinbase Cryptocurrencies                                    2022…
#> 10         86 Commercial Paper                                             2022…
#> 11         72 Daily Treasury Inflation-Indexed Securities                  2022…
#> 12        279 Economic Policy Uncertainty                                  2022…
#> 13        502 Euro Short Term Rate                                         2022…
#> 14        378 Federal Funds Data                                           2022…
#> 15        101 FOMC Press Release                                           2022…
#> 16         18 H.15 Selected Interest Rates                                 2022…
#> 17         22 H.8 Assets and Liabilities of Commercial Banks in the Unite… 2022…
#> 18        500 Inflation Expectations                                       2022…
#> 19        185 Interest Rate on Reserve Balances                            2022…
#> 20        304 Interest Rate Spreads                                        2022…