Search for a FRED series.

Search FRED for a series. Searches can be performed by the text of the series (fredr_search), by series ID (fredr_search_id), by series tags (fredr_search_tags), or related tags (fredr_search_rel_tags).

fredr_search(search_text = NULL, ...)

fredr_search_id(search_text = NULL, ...)

fredr_search_tags(series_search_text = NULL, ...)

fredr_search_rel_tags(series_search_text = NULL, tag_names = NULL, ...)

Arguments

search_text
A string containing the search text. For use with fredr_search and fredr_search_id. See Details for more information.

...
A series of named parameters to be used in the query. Must be of the form param_key = "param_value". Acceptable parameters are endpoint-specific. See fredr_endpoints for a list of endpoints and fredr_docs access to the endpoint web documentation.

series_search_text
A string containing the search text. For use with fredr_search_tags and fredr_search_rel_tags. See Details for more information.

tag_names
A string. Note that this value must be a valid FRED series tag (e.g. 30-year, usa, quarterly, etc.).

Value

A data frame.

Details

Search by various series characteristics:

fredr_searchGet economic data series that match search text (by full text of series).

fredr_search_idGet economic data series that match search text (by series ID).

fredr_search_tagsGet the FRED tags for a series search.

fredr_search_rel_tagsGet the related FRED tags for one or more FRED tags matching a series search. See below for more information about searching for related tags.

From the FRED API documentation on series/search/related_tags: FRED tags are attributes assigned to series. For this request, related FRED tags are the tags assigned to series that match all tags in the tag_names parameter, no tags in the exclude_tag_names parameter, and the search words set by the series_search_text parameter.

References

See FRED API documentation on searching related tags.

See also

Examples

# NOT RUN {
library(fredr)
fredr_search("oil", order_by = "popularity")
fredr_search_id("UNRATE", order_by = "popularity")
fredr_search_tags("gnp")
fredr_search_rel_tags("gnp", "usa")
# }