Return a FRED series as an xts object.

Given a series ID, return observations of that series as an xts object.

fredr_series(series_id, ...)

Arguments

series_id
A string ID for the FRED series. Use fredr_search to find the series_id of a series.

...
A series of named parameters to be used in the query. Must be of the form param_key = "param_value". Acceptable parameters are endpoint-specific. See fredr_endpoints for a list of endpoints and fredr_docs access to the endpoint web documentation.

Value

An xts object with observation dates in the index corresponding to parsed numeric observations as values.

Examples

fredr_series(
  series_id = "UNRATE",
  observation_start = "1980-01-01",
  observation_end = "2000-01-01",
  unit = "chg"
)

#>            UNRATE
#> 1980-01-01    6.3
#> 1980-02-01    6.3
#> 1980-03-01    6.3
#> 1980-04-01    6.9
#> 1980-05-01    7.5
#> 1980-06-01    7.6
#> 1980-07-01    7.8
#> 1980-08-01    7.7
#> 1980-09-01    7.5
#> 1980-10-01    7.5
#> 1980-11-01    7.5
#> 1980-12-01    7.2
#> 1981-01-01    7.5
#> 1981-02-01    7.4
#> 1981-03-01    7.4
#> 1981-04-01    7.2
#> 1981-05-01    7.5
#> 1981-06-01    7.5
#> 1981-07-01    7.2
#> 1981-08-01    7.4
#> 1981-09-01    7.6
#> 1981-10-01    7.9
#> 1981-11-01    8.3
#> 1981-12-01    8.5
#> 1982-01-01    8.6
#> 1982-02-01    8.9
#> 1982-03-01    9.0
#> 1982-04-01    9.3
#> 1982-05-01    9.4
#> 1982-06-01    9.6
#> 1982-07-01    9.8
#> 1982-08-01    9.8
#> 1982-09-01   10.1
#> 1982-10-01   10.4
#> 1982-11-01   10.8
#> 1982-12-01   10.8
#> 1983-01-01   10.4
#> 1983-02-01   10.4
#> 1983-03-01   10.3
#> 1983-04-01   10.2
#> 1983-05-01   10.1
#> 1983-06-01   10.1
#> 1983-07-01    9.4
#> 1983-08-01    9.5
#> 1983-09-01    9.2
#> 1983-10-01    8.8
#> 1983-11-01    8.5
#> 1983-12-01    8.3
#> 1984-01-01    8.0
#> 1984-02-01    7.8
#> 1984-03-01    7.8
#> 1984-04-01    7.7
#> 1984-05-01    7.4
#> 1984-06-01    7.2
#> 1984-07-01    7.5
#> 1984-08-01    7.5
#> 1984-09-01    7.3
#> 1984-10-01    7.4
#> 1984-11-01    7.2
#> 1984-12-01    7.3
#> 1985-01-01    7.3
#> 1985-02-01    7.2
#> 1985-03-01    7.2
#> 1985-04-01    7.3
#> 1985-05-01    7.2
#> 1985-06-01    7.4
#> 1985-07-01    7.4
#> 1985-08-01    7.1
#> 1985-09-01    7.1
#> 1985-10-01    7.1
#> 1985-11-01    7.0
#> 1985-12-01    7.0
#> 1986-01-01    6.7
#> 1986-02-01    7.2
#> 1986-03-01    7.2
#> 1986-04-01    7.1
#> 1986-05-01    7.2
#> 1986-06-01    7.2
#> 1986-07-01    7.0
#> 1986-08-01    6.9
#> 1986-09-01    7.0
#> 1986-10-01    7.0
#> 1986-11-01    6.9
#> 1986-12-01    6.6
#> 1987-01-01    6.6
#> 1987-02-01    6.6
#> 1987-03-01    6.6
#> 1987-04-01    6.3
#> 1987-05-01    6.3
#> 1987-06-01    6.2
#> 1987-07-01    6.1
#> 1987-08-01    6.0
#> 1987-09-01    5.9
#> 1987-10-01    6.0
#> 1987-11-01    5.8
#> 1987-12-01    5.7
#> 1988-01-01    5.7
#> 1988-02-01    5.7
#> 1988-03-01    5.7
#> 1988-04-01    5.4
#> 1988-05-01    5.6
#> 1988-06-01    5.4
#> 1988-07-01    5.4
#> 1988-08-01    5.6
#> 1988-09-01    5.4
#> 1988-10-01    5.4
#> 1988-11-01    5.3
#> 1988-12-01    5.3
#> 1989-01-01    5.4
#> 1989-02-01    5.2
#> 1989-03-01    5.0
#> 1989-04-01    5.2
#> 1989-05-01    5.2
#> 1989-06-01    5.3
#> 1989-07-01    5.2
#> 1989-08-01    5.2
#> 1989-09-01    5.3
#> 1989-10-01    5.3
#> 1989-11-01    5.4
#> 1989-12-01    5.4
#> 1990-01-01    5.4
#> 1990-02-01    5.3
#> 1990-03-01    5.2
#> 1990-04-01    5.4
#> 1990-05-01    5.4
#> 1990-06-01    5.2
#> 1990-07-01    5.5
#> 1990-08-01    5.7
#> 1990-09-01    5.9
#> 1990-10-01    5.9
#> 1990-11-01    6.2
#> 1990-12-01    6.3
#> 1991-01-01    6.4
#> 1991-02-01    6.6
#> 1991-03-01    6.8
#> 1991-04-01    6.7
#> 1991-05-01    6.9
#> 1991-06-01    6.9
#> 1991-07-01    6.8
#> 1991-08-01    6.9
#> 1991-09-01    6.9
#> 1991-10-01    7.0
#> 1991-11-01    7.0
#> 1991-12-01    7.3
#> 1992-01-01    7.3
#> 1992-02-01    7.4
#> 1992-03-01    7.4
#> 1992-04-01    7.4
#> 1992-05-01    7.6
#> 1992-06-01    7.8
#> 1992-07-01    7.7
#> 1992-08-01    7.6
#> 1992-09-01    7.6
#> 1992-10-01    7.3
#> 1992-11-01    7.4
#> 1992-12-01    7.4
#> 1993-01-01    7.3
#> 1993-02-01    7.1
#> 1993-03-01    7.0
#> 1993-04-01    7.1
#> 1993-05-01    7.1
#> 1993-06-01    7.0
#> 1993-07-01    6.9
#> 1993-08-01    6.8
#> 1993-09-01    6.7
#> 1993-10-01    6.8
#> 1993-11-01    6.6
#> 1993-12-01    6.5
#> 1994-01-01    6.6
#> 1994-02-01    6.6
#> 1994-03-01    6.5
#> 1994-04-01    6.4
#> 1994-05-01    6.1
#> 1994-06-01    6.1
#> 1994-07-01    6.1
#> 1994-08-01    6.0
#> 1994-09-01    5.9
#> 1994-10-01    5.8
#> 1994-11-01    5.6
#> 1994-12-01    5.5
#> 1995-01-01    5.6
#> 1995-02-01    5.4
#> 1995-03-01    5.4
#> 1995-04-01    5.8
#> 1995-05-01    5.6
#> 1995-06-01    5.6
#> 1995-07-01    5.7
#> 1995-08-01    5.7
#> 1995-09-01    5.6
#> 1995-10-01    5.5
#> 1995-11-01    5.6
#> 1995-12-01    5.6
#> 1996-01-01    5.6
#> 1996-02-01    5.5
#> 1996-03-01    5.5
#> 1996-04-01    5.6
#> 1996-05-01    5.6
#> 1996-06-01    5.3
#> 1996-07-01    5.5
#> 1996-08-01    5.1
#> 1996-09-01    5.2
#> 1996-10-01    5.2
#> 1996-11-01    5.4
#> 1996-12-01    5.4
#> 1997-01-01    5.3
#> 1997-02-01    5.2
#> 1997-03-01    5.2
#> 1997-04-01    5.1
#> 1997-05-01    4.9
#> 1997-06-01    5.0
#> 1997-07-01    4.9
#> 1997-08-01    4.8
#> 1997-09-01    4.9
#> 1997-10-01    4.7
#> 1997-11-01    4.6
#> 1997-12-01    4.7
#> 1998-01-01    4.6
#> 1998-02-01    4.6
#> 1998-03-01    4.7
#> 1998-04-01    4.3
#> 1998-05-01    4.4
#> 1998-06-01    4.5
#> 1998-07-01    4.5
#> 1998-08-01    4.5
#> 1998-09-01    4.6
#> 1998-10-01    4.5
#> 1998-11-01    4.4
#> 1998-12-01    4.4
#> 1999-01-01    4.3
#> 1999-02-01    4.4
#> 1999-03-01    4.2
#> 1999-04-01    4.3
#> 1999-05-01    4.2
#> 1999-06-01    4.3
#> 1999-07-01    4.3
#> 1999-08-01    4.2
#> 1999-09-01    4.2
#> 1999-10-01    4.1
#> 1999-11-01    4.1
#> 1999-12-01    4.0
#> 2000-01-01    4.0

fredr_series(
  series_id = "OILPRICE",
  frequency = "q",
  aggregation_method = "avg",
  unit = "log"
)

#>            OILPRICE
#> 1946-01-01    1.170
#> 1946-04-01    1.270
#> 1946-07-01    1.437
#> 1946-10-01    1.553
#> 1947-01-01    1.620
#> 1947-04-01    1.870
#> 1947-07-01    1.870
#> 1947-10-01    2.003
#> 1948-01-01    2.570
#> 1948-04-01    2.570
#> 1948-07-01    2.570
#> 1948-10-01    2.570
#> 1949-01-01    2.570
#> 1949-04-01    2.570
#> 1949-07-01    2.570
#> 1949-10-01    2.570
#> 1950-01-01    2.570
#> 1950-04-01    2.570
#> 1950-07-01    2.570
#> 1950-10-01    2.570
#> 1951-01-01    2.570
#> 1951-04-01    2.570
#> 1951-07-01    2.570
#> 1951-10-01    2.570
#> 1952-01-01    2.570
#> 1952-04-01    2.570
#> 1952-07-01    2.570
#> 1952-10-01    2.570
#> 1953-01-01    2.570
#> 1953-04-01    2.653
#> 1953-07-01    2.820
#> 1953-10-01    2.820
#> 1954-01-01    2.820
#> 1954-04-01    2.820
#> 1954-07-01    2.820
#> 1954-10-01    2.820
#> 1955-01-01    2.820
#> 1955-04-01    2.820
#> 1955-07-01    2.820
#> 1955-10-01    2.820
#> 1956-01-01    2.820
#> 1956-04-01    2.820
#> 1956-07-01    2.820
#> 1956-10-01    2.820
#> 1957-01-01    2.987
#> 1957-04-01    3.070
#> 1957-07-01    3.070
#> 1957-10-01    3.047
#> 1958-01-01    3.070
#> 1958-04-01    3.070
#> 1958-07-01    3.070
#> 1958-10-01    3.023
#> 1959-01-01    2.990
#> 1959-04-01    2.970
#> 1959-07-01    2.970
#> 1959-10-01    2.970
#> 1960-01-01    2.970
#> 1960-04-01    2.970
#> 1960-07-01    2.970
#> 1960-10-01    2.970
#> 1961-01-01    2.970
#> 1961-04-01    2.970
#> 1961-07-01    2.970
#> 1961-10-01    2.970
#> 1962-01-01    2.970
#> 1962-04-01    2.970
#> 1962-07-01    2.970
#> 1962-10-01    2.970
#> 1963-01-01    2.970
#> 1963-04-01    2.970
#> 1963-07-01    2.970
#> 1963-10-01    2.970
#> 1964-01-01    2.970
#> 1964-04-01    2.970
#> 1964-07-01    2.920
#> 1964-10-01    2.920
#> 1965-01-01    2.920
#> 1965-04-01    2.920
#> 1965-07-01    2.920
#> 1965-10-01    2.920
#> 1966-01-01    2.920
#> 1966-04-01    2.920
#> 1966-07-01    2.937
#> 1966-10-01    2.970
#> 1967-01-01    2.990
#> 1967-04-01    3.000
#> 1967-07-01    3.047
#> 1967-10-01    3.070
#> 1968-01-01    3.070
#> 1968-04-01    3.070
#> 1968-07-01    3.070
#> 1968-10-01    3.070
#> 1969-01-01    3.130
#> 1969-04-01    3.350
#> 1969-07-01    3.350
#> 1969-10-01    3.350
#> 1970-01-01    3.350
#> 1970-04-01    3.350
#> 1970-07-01    3.310
#> 1970-10-01    3.393
#> 1971-01-01    3.560
#> 1971-04-01    3.560
#> 1971-07-01    3.560
#> 1971-10-01    3.560
#> 1972-01-01    3.560
#> 1972-04-01    3.560
#> 1972-07-01    3.560
#> 1972-10-01    3.560
#> 1973-01-01    3.560
#> 1973-04-01    3.560
#> 1973-07-01    4.060
#> 1973-10-01    4.310
#> 1974-01-01   10.110
#> 1974-04-01   10.110
#> 1974-07-01   10.110
#> 1974-10-01   11.160
#> 1975-01-01   11.160
#> 1975-04-01   11.160
#> 1975-07-01   11.160
#> 1975-10-01   11.160
#> 1976-01-01   11.763
#> 1976-04-01   12.170
#> 1976-07-01   12.747
#> 1976-10-01   13.900
#> 1977-01-01   13.900
#> 1977-04-01   13.900
#> 1977-07-01   14.533
#> 1977-10-01   14.850
#> 1978-01-01   14.850
#> 1978-04-01   14.850
#> 1978-07-01   14.850
#> 1978-10-01   14.850
#> 1979-01-01   15.517
#> 1979-04-01   17.683
#> 1979-07-01   25.583
#> 1979-10-01   30.833
#> 1980-01-01   35.833
#> 1980-04-01   39.500
#> 1980-07-01   37.833
#> 1980-10-01   36.333
#> 1981-01-01   38.000
#> 1981-04-01   37.333
#> 1981-07-01   36.000
#> 1981-10-01   35.333
#> 1982-01-01   31.297
#> 1982-04-01   34.817
#> 1982-07-01   34.580
#> 1982-10-01   33.850
#> 1983-01-01   29.653
#> 1983-04-01   30.537
#> 1983-07-01   31.560
#> 1983-10-01   29.830
#> 1984-01-01   30.199
#> 1984-04-01   30.368
#> 1984-07-01   29.104
#> 1984-10-01   27.431
#> 1985-01-01   27.050
#> 1985-04-01   27.857
#> 1985-07-01   27.791
#> 1985-10-01   29.193
#> 1986-01-01   17.002
#> 1986-04-01   13.920
#> 1986-07-01   13.860
#> 1986-10-01   15.378
#> 1987-01-01   18.229
#> 1987-04-01   19.364
#> 1987-07-01   20.385
#> 1987-10-01   18.670
#> 1988-01-01   16.714
#> 1988-04-01   17.281
#> 1988-07-01   15.163
#> 1988-10-01   14.680
#> 1989-01-01   18.418
#> 1989-04-01   20.359
#> 1989-07-01   19.253
#> 1989-10-01   20.333
#> 1990-01-01   21.722
#> 1990-04-01   17.897
#> 1990-07-01   26.500
#> 1990-10-01   31.853
#> 1991-01-01   21.781
#> 1991-04-01   20.753
#> 1991-07-01   21.655
#> 1991-10-01   21.737
#> 1992-01-01   18.910
#> 1992-04-01   21.186
#> 1992-07-01   21.670
#> 1992-10-01   20.479
#> 1993-01-01   19.825
#> 1993-04-01   19.760
#> 1993-07-01   17.796
#> 1993-10-01   16.451
#> 1994-01-01   14.813
#> 1994-04-01   17.777
#> 1994-07-01   18.497
#> 1994-10-01   17.657
#> 1995-01-01   18.357
#> 1995-04-01   19.343
#> 1995-07-01   17.853
#> 1995-10-01   18.157
#> 1996-01-01   19.770
#> 1996-04-01   21.757
#> 1996-07-01   22.423
#> 1996-10-01   24.667
#> 1997-01-01   22.790
#> 1997-04-01   19.907
#> 1997-07-01   19.783
#> 1997-10-01   19.917
#> 1998-01-01   15.930
#> 1998-04-01   14.653
#> 1998-07-01   14.130
#> 1998-10-01   12.840
#> 1999-01-01   13.047
#> 1999-04-01   17.660
#> 1999-07-01   21.737
#> 1999-10-01   24.563
#> 2000-01-01   28.807
#> 2000-04-01   28.783
#> 2000-07-01   31.623
#> 2000-10-01   31.980
#> 2001-01-01   28.810
#> 2001-04-01   27.883
#> 2001-07-01   26.600
#> 2001-10-01   20.403
#> 2002-01-01   21.610
#> 2002-04-01   26.270
#> 2002-07-01   28.330
#> 2002-10-01   28.180
#> 2003-01-01   34.120
#> 2003-04-01   29.037
#> 2003-07-01   30.213
#> 2003-10-01   31.190
#> 2004-01-01   35.257
#> 2004-04-01   38.330
#> 2004-07-01   43.860
#> 2004-10-01   48.307
#> 2005-01-01   49.707
#> 2005-04-01   53.043
#> 2005-07-01   63.080
#> 2005-10-01   60.033
#> 2006-01-01   63.347
#> 2006-04-01   70.530
#> 2006-07-01   70.443
#> 2006-10-01   60.093
#> 2007-01-01   58.130
#> 2007-04-01   64.970
#> 2007-07-01   75.500
#> 2007-10-01   90.850
#> 2008-01-01   97.953
#> 2008-04-01  123.963
#> 2008-07-01  117.983
#> 2008-10-01   58.370
#> 2009-01-01   42.960
#> 2009-04-01   59.543
#> 2009-07-01   68.203
#> 2009-10-01   76.067
#> 2010-01-01   78.627
#> 2010-04-01   77.890
#> 2010-07-01   76.167
#> 2010-10-01   85.027
#> 2011-01-01   93.980
#> 2011-04-01  102.553
#> 2011-07-01   89.710
#> 2011-10-01   94.063
#> 2012-01-01  102.893
#> 2012-04-01   93.480
#> 2012-07-01   92.270
#> 2012-10-01   88.160
#> 2013-01-01   94.353
#> 2013-04-01   94.223
#> 2013-07-01       NA