Get the FRED tags for a series search.

Return the FRED tags by searching for matches in series text.

fredr_series_search_tags(
  series_search_text,
  ...,
  tag_names = NULL,
  tag_group_id = NULL,
  tag_search_text = NULL,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

series_search_text

A string containing the series search text.

...

These dots only exist for future extensions and should be empty.

tag_names

A semicolon delimited string of tag names to return. Defaults no filtering by tag names.

tag_group_id

A string indicating the tag group id to filter tags by type. Defaults to no filtering by tag group. Possible values are

"freq" = Frequency

"gen" = General or Concept

"geo" = Geography

"geot" = Geography Type

"rls" = Release

"seas" = Seasonal Adjustment

"src" = Source

tag_search_text

A string to match tag names. Defaults to no filtering by tag name matching.

limit

An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset

An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by

A string indicating the attribute to order results by. Defaults to "series_count". Possible values are:

"series_count"

"popularity"

"created"

"name"

"group_id"

sort_order

A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start

A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end

A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object where each row represents a series tag matching the query. Data include the tag name, group ID, tag creation date, popularity, series count, and additional notes.

References

API Documentation:

series/search/tags

See also

Examples

if (fredr_has_key()) {
# Search for tags matching the series text "gnp"
fredr_series_search_tags("gnp")
# Search for tags matching the series text "oil" and the tag text "usa"
fredr_series_search_tags(
  series_search_text = "oil",
  tag_search_text = "usa"
)
# Search for tags matching the series text "oil" and the tag text "usa".
# Return only results in the "geo" (Geography) group
fredr_series_search_tags(
  series_search_text = "oil",
  tag_group_id = "geo",
  tag_search_text = "usa"
)
}
#> # A tibble: 2 × 6
#>   name  group_id notes                    created        popularity series_count
#>   <chr> <chr>    <chr>                    <chr>               <int>        <int>
#> 1 usa   geo      United States of America 2012-02-27 10…        100         5314
#> 2 ga    geo      Georgia (U.S. state)     2012-02-27 10…         49          192