Get the tags for a FRED series

Given a series ID, return associated tags for the series as a tibble object.

fredr_series_tags(
  series_id,
  ...,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

series_id
A string ID for the FRED series.

...
These dots only exist for future extensions and should be empty.

order_by
A string indicating the attribute by which to order the Possible values include "series_count" (default), "popularity", "created", "name", and "group_id".

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble`` object where each row is represents a tag associated with the series specified by series_id`. Data include tag name, group ID, popularity, series count, tag creation date, and additional notes.

API Documentation

fred/series/tags

See also

Examples

if (fredr_has_key()) {
# Return all tags assigned to the "UNRATE" series and order the results by
# group ID.
fredr_series_tags(series_id = "UNRATE", order_by = "group_id")
}

#> # A tibble: 14 x 6
#>    name           group_id notes             created     popularity series_count
#>    <chr>          <chr>    <chr>             <chr>            <int>        <int>
#>  1 public domain… cc        NA               2018-12-17…        100       583962
#>  2 monthly        freq     ""                2012-02-27…         93       213346
#>  3 rate           gen      ""                2012-02-27…         87        47336
#>  4 unemployment   gen      ""                2012-02-27…         78        32836
#>  5 household sur… gen      "Current Populat… 2012-06-18…         65        21530
#>  6 labor          gen      ""                2012-02-27…         61        33498
#>  7 16 years +     gen      ""                2012-02-27…         59         5802
#>  8 civilian       gen      ""                2012-02-27…         55         9578
#>  9 headline figu… gen      ""                2013-11-19…         53           58
#> 10 labor underut… gen      ""                2012-05-10…         37          718
#> 11 usa            geo      "United States o… 2012-02-27…        100       652060
#> 12 nation         geot     ""                2012-02-27…         99       257804
#> 13 sa             seas     "Seasonally Adju… 2012-02-27…         88        81514
#> 14 bls            src      "Bureau of Labor… 2012-02-27…         87       175670