Get the data vintage dates for a FRED series

Given a series ID, return a sequence of dates in history when a series' data values were revised or new data values were released as a tibble object.

fredr_series_vintagedates(series_id = NULL, limit = NULL, offset = NULL,
  sort_order = NULL, realtime_start = NULL, realtime_end = NULL)

Arguments

series_id
A string ID for the FRED series. Required parameter.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object where each row is a distinct vintage date.

API Documentation

fred/series/vintagedates

See also

Examples

# All data vintages for the "UNRATE" series
fredr_series_vintagedates(series_id = "UNRATE")

#> # A tibble: 701 x 1
#>    vintage_dates
#>    <chr>        
#>  1 1960-03-15   
#>  2 1960-04-12   
#>  3 1960-05-11   
#>  4 1960-06-13   
#>  5 1960-07-12   
#>  6 1960-08-10   
#>  7 1960-09-09   
#>  8 1960-10-10   
#>  9 1960-11-10   
#> 10 1960-12-12   
#> # ... with 691 more rows

# 10 most recent data vintages for the "UNRATE" series
fredr_series_vintagedates(series_id = "UNRATE", limit = 10L, sort_order = "desc")

#> # A tibble: 10 x 1
#>    vintage_dates
#>    <chr>        
#>  1 2018-07-06   
#>  2 2018-06-01   
#>  3 2018-05-04   
#>  4 2018-04-06   
#>  5 2018-03-09   
#>  6 2018-02-02   
#>  7 2018-01-05   
#>  8 2017-12-08   
#>  9 2017-11-03   
#> 10 2017-10-06