Get the releases for a source

Get the releases for a source

fredr_source_releases(
  source_id,
  ...,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

source_id
An integer ID for the data source.

...
These dots only exist for future extensions and should be empty.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
A string indicating which attribute should be used to order the results. Possible values:

"release_id" (default)

"name"

"press_release"

"realtime_start"

"realtime_end"

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/source/releases

See also

Examples

if (fredr_has_key()) {
# Board of Governors
fredr_source_releases(source_id = 1L)

# University of Michigan
fredr_source_releases(source_id = 14L, realtime_start = as.Date("1950-01-01"))
}

#> # A tibble: 1 x 6
#>      id realtime_start realtime_end name          press_release link            
#>   <int> <chr>          <chr>        <chr>         <lgl>         <chr>           
#> 1    91 1998-07-31     9999-12-31   Surveys of C… TRUE          http://www.sca.…