Get all sources of economic data

Get all sources of economic data

fredr_sources(
  ...,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

...
These dots only exist for future extensions and should be empty.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
A string indicating which attribute should be used to order the results. Possible values: "source_id" (default), "name", "realtime_start", "realtime_end".

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/sources

See also

Examples

if (fredr_has_key()) {
fredr_sources(limit = 20L)
}

#> # A tibble: 20 x 6
#>       id realtime_start realtime_end name           link               notes    
#>    <int> <chr>          <chr>        <chr>          <chr>              <chr>    
#>  1     1 2021-02-15     2021-02-15   Board of Gove… http://www.federa… NA       
#>  2     3 2021-02-15     2021-02-15   Federal Reser… https://www.phila… NA       
#>  3     4 2021-02-15     2021-02-15   Federal Reser… http://www.stloui… NA       
#>  4     6 2021-02-15     2021-02-15   Federal Finan… http://www.ffiec.… NA       
#>  5    11 2021-02-15     2021-02-15   Dow Jones & C… http://www.dowjon… NA       
#>  6    14 2021-02-15     2021-02-15   University of… https://www.umich… NA       
#>  7    15 2021-02-15     2021-02-15   Council of Ec… https://www.white… NA       
#>  8    16 2021-02-15     2021-02-15   U.S. Office o… https://www.white… NA       
#>  9    17 2021-02-15     2021-02-15   U.S. Congress… http://www.cbo.go… NA       
#> 10    18 2021-02-15     2021-02-15   U.S. Bureau o… http://www.bea.go… NA       
#> 11    19 2021-02-15     2021-02-15   U.S. Census B… http://www.census… NA       
#> 12    21 2021-02-15     2021-02-15   U.S. Departme… http://www.hud.go… NA       
#> 13    22 2021-02-15     2021-02-15   U.S. Bureau o… https://www.bls.g… NA       
#> 14    23 2021-02-15     2021-02-15   U.S. Departme… http://www.fiscal… NA       
#> 15    26 2021-02-15     2021-02-15   Haver Analyti… http://www.haver.… NA       
#> 16    31 2021-02-15     2021-02-15   Reserve Bank … http://www.rba.go… NA       
#> 17    32 2021-02-15     2021-02-15   Deutsche Bund… http://www.bundes… NA       
#> 18    33 2021-02-15     2021-02-15   Bank of Italy  http://www.bancad… Banca d'…
#> 19    34 2021-02-15     2021-02-15   Swiss Nationa… http://www.snb.ch/ NA       
#> 20    35 2021-02-15     2021-02-15   Central Bank … http://www.tcmb.g… Türkiye…