Get FRED series tags

Get FRED tags. Optionally, filter results by tag name, tag group, or search text. FRED tags are attributes assigned to a series. By default, all tags are returned, unfiltered, up to the limit.

fredr_tags(tag_names = NULL, tag_group_id = NULL, search_text = NULL,
  limit = NULL, offset = NULL, order_by = NULL, sort_order = NULL,
  realtime_start = NULL, realtime_end = NULL)

Arguments

tag_names
A semicolon delimited string of tag names to only include in the response. No filtering by tag names by default (i.e. all FRED tags returned).

tag_group_id
A string tag group id to filter tags by type. No filtering by tag group by default. Possible values are:

"freq" = Frequency

"gen" = General or Concept

"geo" = Geography

"geot" = Geography Type

"rls" = Release

"seas" = Seasonal Adjustment

"src" = Source

search_text
A string indicating the words to find matching tags with. No filtering by search words by default.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
Order results by values of the specified attribute. Possible values are:

"series_count" (default)

"popularity"

"created"

"name"

"group_id"

sort_order
A string representing the order of the resulting series, sorted by the attribute values specified by order_by. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble containing tags and their descriptions. Data include tag name, group ID, popularity, series count, tag creation date, and additional notes.

API Documentation

fred/tags

See also

Examples

# \donttest{
# Information for all tags
fredr_tags()

#> # A tibble: 1,000 x 6
#>    name             group_id notes          created      popularity series_count
#>    <chr>            <chr>    <chr>          <chr>             <int>        <int>
#>  1 nsa              seas     Not Seasonall… 2012-02-27 …        100       514570
#>  2 public domain: … cc       NA             2018-12-17 …        100       471185
#>  3 usa              geo      United States… 2012-02-27 …         99       429976
#>  4 annual           freq     ""             2012-02-27 …         91       362462
#>  5 nation           geot     ""             2012-02-27 …        100       215998
#>  6 county           geot     County or Cou… 2012-02-27 …         74       196925
#>  7 census           src      Census         2012-02-27 …         77       182831
#>  8 bls              src      Bureau of Lab… 2012-02-27 …         85       136652
#>  9 monthly          freq     ""             2012-02-27 …         92       120445
#> 10 copyrighted: ci… cc       NA             2018-12-17 …         88       115744
#> # … with 990 more rows

# Information for just the "gdp" and "oecd" tags
fredr_tags(tag_names = "gdp;oecd")

#> # A tibble: 2 x 6
#>   name  group_id notes                     created       popularity series_count
#>   <chr> <chr>    <chr>                     <chr>              <int>        <int>
#> 1 oecd  src      Org. for Economic Co-ope… 2012-02-27 1…         77        63847
#> 2 gdp   gen      Gross Domestic Product    2012-02-27 1…         83        49302

# Information for all tags in the "geo" group
fredr_tags(tag_group_id = "geo")

#> # A tibble: 1,000 x 6
#>    name  group_id notes               created            popularity series_count
#>    <chr> <chr>    <chr>               <chr>                   <int>        <int>
#>  1 usa   geo      United States of A… 2012-02-27 10:18:…         99       429976
#>  2 tx    geo      Texas               2012-02-27 10:18:…         55        23576
#>  3 ga    geo      Georgia (U.S. stat… 2012-02-27 10:18:…         46        13101
#>  4 ca    geo      California          2012-02-27 10:18:…         59        12683
#>  5 va    geo      Virginia            2012-02-27 10:18:…         45        11270
#>  6 il    geo      Illinois            2012-02-27 10:18:…         47        10617
#>  7 ky    geo      Kentucky            2012-02-27 10:18:…         43        10246
#>  8 mo    geo      Missouri            2012-02-27 10:18:…         44        10154
#>  9 nc    geo      North Carolina      2012-02-27 10:18:…         48         9931
#> 10 in    geo      Indiana             2012-02-27 10:18:…         46         9799
#> # … with 990 more rows

# Information for tags matching the text "unemployment"
fredr_tags(search_text = "unemployment")

#> # A tibble: 2 x 6
#>   name     group_id notes                  created       popularity series_count
#>   <chr>    <chr>    <chr>                  <chr>              <int>        <int>
#> 1 unemplo… gen      ""                     2012-02-27 1…         75        32202
#> 2 nairu    gen      Non-Accelerating Infl… 2012-08-10 1…         18            2

# }