Find FRED series matching tag names

Get the series matching tags in the tag_names parameter. Exclude tags in the exclude_tag_names parameter.

fredr_tags_series(
  tag_names,
  ...,
  exclude_tag_names = NULL,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

tag_names
A semicolon delimited string of tag names to find series using.

...
These dots only exist for future extensions and should be empty.

exclude_tag_names
A semicolon delimited string of tag names that series match none of. No exclusions are done by default.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
A string indicating which attribute by which to order the results of the query. Possible values include:

"series_id" (default)

"title"

"units"

"frequency"

"seasonal_adjustment"

"realtime_start"

"realtime_end"

"last_updated"

"observation_start"

"observation_end"

"popularity"

"group_popularity"

sort_order
A string representing the order of the resulting series, sorted by the attribute values specified by order_by. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object containing FRED series with tags matching tag_names and their descriptions.

API Documentation

fred/tags/series

See also

Examples

if (fredr_has_key()) {
# All series tagged with "gdp"
fredr_tags_series(tag_names = "gdp")
# All series tagged with "gdp" and not tagged with "quarterly"
fredr_tags_series(
   tag_names = "gdp",
   exclude_tag_names = "quarterly"
 )
# Top 100 most popular non-quarterly series matching GDP
fredr_tags_series(
   tag_names = "gdp",
   exclude_tag_names = "quarterly",
   order_by = "popularity",
   limit = 100L
)
}

#> # A tibble: 100 x 16
#>    id    realtime_start realtime_end title observation_sta… observation_end
#>    <chr> <chr>          <chr>        <chr> <chr>            <chr>          
#>  1 H407… 2020-12-14     2020-12-14   Comp… 1929-01-01       1948-01-01     
#>  2 B407… 2020-12-14     2020-12-14   Comp… 1987-01-01       2000-01-01     
#>  3 J404… 2020-12-14     2020-12-14   Comp… 1948-01-01       1987-01-01     
#>  4 N404… 2020-12-14     2020-12-14   Comp… 1998-01-01       2019-01-01     
#>  5 B405… 2020-12-14     2020-12-14   Comp… 1987-01-01       2000-01-01     
#>  6 H405… 2020-12-14     2020-12-14   Comp… 1929-01-01       1948-01-01     
#>  7 J405… 2020-12-14     2020-12-14   Comp… 1948-01-01       1969-01-01     
#>  8 B490… 2020-12-14     2020-12-14   Empl… 1987-01-01       2000-01-01     
#>  9 J491… 2020-12-14     2020-12-14   Empl… 1948-01-01       1987-01-01     
#> 10 B480… 2020-12-14     2020-12-14   Empl… 1948-01-01       1987-01-01     
#> # … with 90 more rows, and 10 more variables: frequency <chr>,
#> #   frequency_short <chr>, units <chr>, units_short <chr>,
#> #   seasonal_adjustment <chr>, seasonal_adjustment_short <chr>,
#> #   last_updated <chr>, popularity <int>, group_popularity <int>, notes <chr>