Send a request to the FRED API

Send a request to the FRED API

fredr(endpoint, ..., to_frame = TRUE, print_req = FALSE)

Arguments

endpoint
A character vector of length 1 representing the FRED API path of interest. See https://api.stlouisfed.org/docs/fred/ for a list of endpoint possible values.

...
A series of paramters to be used in the query. Of the form param_key = 'param_value'. Acceptable parameters are specific to the API path. See https://api.stlouisfed.org/docs/fred/ for a list of recognized parameters for each endpoint.

to_frame
A boolean value indicating whether or not the response should be parsed and formatted as a data.frame. Default is TRUE.

print_req
A boolean value indicating whether or not the request should be printed as well. Useful for debugging. Default is FALSE.

Value

If to_frame = TRUE, a data.frame containing the parsed response. If to_frame = FALSE, an object returned directly from httr::GET().

Examples

 fredr(endpoint = "series/observations",
       series_id = "GNPCA",
       observation_start = "1990-01-01",
       observation_end = "2000-01-01")

#> # A tibble: 11 x 4
#>    realtime_start realtime_end       date   value
#>             <chr>        <chr>      <chr>   <chr>
#>  1     2017-06-03   2017-06-03 1990-01-01  9012.6
#>  2     2017-06-03   2017-06-03 1991-01-01  8998.1
#>  3     2017-06-03   2017-06-03 1992-01-01  9314.9
#>  4     2017-06-03   2017-06-03 1993-01-01  9570.0
#>  5     2017-06-03   2017-06-03 1994-01-01  9942.9
#>  6     2017-06-03   2017-06-03 1995-01-01 10217.3
#>  7     2017-06-03   2017-06-03 1996-01-01 10606.5
#>  8     2017-06-03   2017-06-03 1997-01-01 11070.1
#>  9     2017-06-03   2017-06-03 1998-01-01 11553.1
#> 10     2017-06-03   2017-06-03 1999-01-01 12103.4
#> 11     2017-06-03   2017-06-03 2000-01-01 12608.8

 not_run({
   library(httr)
   library(tidyverse)
   resp <- fredr(endpoint = "series/observations",
                 series_id = "GNPCA",
                 observation_start = "1990-01-01",
                 observation_end = "2000-01-01",
                 to_frame = FALSE)
   data <- resp %>%
     httr::content() %>%
     .$observations %>%
     purrr::map_df(.f = function(x) tibble::as_tibble(x = x))
})