Send a request to the FRED API

fredr forms and submits a request to a specified endpoint of the FRED API.

fredr(endpoint, ..., to_frame = TRUE, print_req = FALSE)

Arguments

endpoint
A string representing the FRED API endpoint of interest. See fredr_endpoints for a list of endpoint possible values.

...
A series of named parameters to be used in the query. Must be of the form param_key = "param_value". Acceptable parameters are endpoint-specific. See fredr_endpoints for a list of endpoints and fredr_docs access to the endpoint web documentation.

to_frame
A boolean value indicating whether or not the response should be parsed and formatted as a data frame. If FALSE, a response object is returned and further processing can be done with content. Default is TRUE.

print_req
A boolean value indicating whether or not the request should be printed as well. Useful for debugging. Default is FALSE.

Value

If to_frame = TRUE, a data frame containing the parsed response. If to_frame = FALSE, an object returned directly from GET.

References

https://api.stlouisfed.org/docs/fred/.

See also

Examples

fredr(
  endpoint = "series/observations",
  series_id = "GNPCA",
  observation_start = "1990-01-01",
  observation_end = "2000-01-01"
)

#> # A tibble: 11 x 4
#>    realtime_start realtime_end date       value    
#>  * <chr>          <chr>        <chr>      <chr>    
#>  1 2018-07-08     2018-07-08   1990-01-01 9012.628 
#>  2 2018-07-08     2018-07-08   1991-01-01 8998.131 
#>  3 2018-07-08     2018-07-08   1992-01-01 9314.879 
#>  4 2018-07-08     2018-07-08   1993-01-01 9570.012 
#>  5 2018-07-08     2018-07-08   1994-01-01 9942.936 
#>  6 2018-07-08     2018-07-08   1995-01-01 10217.258
#>  7 2018-07-08     2018-07-08   1996-01-01 10606.548
#>  8 2018-07-08     2018-07-08   1997-01-01 11070.064
#>  9 2018-07-08     2018-07-08   1998-01-01 11553.116
#> 10 2018-07-08     2018-07-08   1999-01-01 12103.387
#> 11 2018-07-08     2018-07-08   2000-01-01 12608.778
# NOT RUN {
# Compare with to_frame = TRUE
library(httr)
library(tidyverse)
resp <- fredr(
  endpoint = "series/observations",
  series_id = "GNPCA",
  observation_start = "1990-01-01",
  observation_end = "2000-01-01",
  to_frame = FALSE
)
data <- resp %>%
httr::content() %>%
  .$observations %>%
  purrr::map_df(.f = function(x) tibble::as_tibble(x = x))
# }