Get the series in a category

Get the series in a category

fredr_category_series(category_id = 0L, filter_variable = NULL,
  filter_value = NULL, tag_names = NULL, exclude_tag_names = NULL,
  limit = NULL, offset = NULL, order_by = NULL, sort_order = NULL,
  realtime_start = NULL, realtime_end = NULL)

Arguments

category_id
An integer ID for the category. Default is 0 for the root category. Required parameter.

filter_variable
A string indicating which attribute to indicate the attribute that results are filtered by. Possible values include: "frequency", "units", "seasonal_adjustment". No filtering by default.

filter_value
A string giving the value of the filter_variable attribute to filter results by. filter_variable must be set. No filtering by default.

tag_names
A string indicating which series tags to match. Multiple tags can be delimited by a semicolon in a single string (e.g. `"usa;gnp"``).

exclude_tag_names
A string indicating which series tags should not be matched. Multiple tags can be delimited by a semicolon in a single string (e.g. `"usa;gnp"``).

limit
An positive integer indicating maximum number of results to return. Possible values are any integer between 1 and 1000 (default), inclusive.

offset
An non-negative integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
A string indicating an attribute by which the results are ordered by. Possible values include:

"series_id" (default)

"title"

"units"

"frequency"

"seasonal_adjustment"

"realtime_start"

"realtime_end"

"last_updated"

"observation_start"

"observation_end"

"popularity"

"group_popularity"

sort_order
A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object with information for series matching the request for the category specified in category_id.

API Documentation

fred/category/series

See also

Examples

# NOT RUN {
# Top 10 most popular series belonging to the "Employment Cost Index" category
fredr_category_series(category_id = 1L, limit = 10L, order_by = "popularity")
# Series in the "Employment Cost Index" category, ordered by descending observation frequency
fredr_category_series(category_id = 4L, order_by = "frequency", sort_order = "desc")
# }