Get related FRED tags given one or more tags

Get related FRED tags. Optionally, filter results by tag group, or search text. Related FRED tags are the tags assigned to series that match all tags in the tag_names parameter and no tags in the exclude_tag_names parameter.

fredr_related_tags(
  tag_names,
  ...,
  exclude_tag_names = NULL,
  tag_group_id = NULL,
  search_text = NULL,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

tag_names
A semicolon delimited string of tag names to be related to.

...
These dots only exist for future extensions and should be empty.

exclude_tag_names
A semicolon delimited string of tag names that series match none of. No exclusions are done by default.

tag_group_id
A string tag group id to filter tags by type. No filtering by tag group by default. Possible values are:

"freq" = Frequency

"gen" = General or Concept

"geo" = Geography

"geot" = Geography Type

"rls" = Release

"seas" = Seasonal Adjustment

"src" = Source

search_text
A string indicating the words to find matching tags with. No filtering by search words by default.

limit
An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset
An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by
Order results by values of the specified attribute. Possible values are:

"series_count" (default)

"popularity"

"created"

"name"

"group_id"

sort_order
A string representing the order of the resulting series, sorted by the attribute values specified by order_by. Possible values are: "asc" (default), and "desc".

realtime_start
A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end
A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble containing tags related to tag_names and their descriptions. Data include tag name, group ID, popularity, series count, tag creation date, and additional notes.

API Documentation

fred/related_tags

See also

Examples

if (fredr_has_key()) {

fredr_related_tags(tag_names = "monetary aggregates;weekly")

fredr_related_tags(
   tag_names = "monetary aggregates;weekly",
   tag_group_id = "gen"
 )

 }

#> # A tibble: 4 x 6
#>   name        group_id notes          created            popularity series_count
#>   <chr>       <chr>    <chr>          <chr>                   <int>        <int>
#> 1 discontinu… gen      ""             2012-02-27 10:18:…         67           10
#> 2 m2          gen      "M2 Money Sto… 2012-02-27 10:18:…         43            6
#> 3 mzm         gen      "MZM Money St… 2012-02-27 10:18:…         27            4
#> 4 m1          gen      "M1 Money Sto… 2012-02-27 10:18:…         47            2