Get the series on a release of economic data

Get the series on a release of economic data

fredr_release_series(
  release_id,
  ...,
  filter_variable = NULL,
  filter_value = NULL,
  tag_names = NULL,
  exclude_tag_names = NULL,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

release_id

An integer ID of the release.

...

These dots only exist for future extensions and should be empty.

filter_variable

A string indicating which attribute to indicate the attribute that results are filtered by. Possible values include: "frequency", "units", "seasonal_adjustment". No filtering by default.

filter_value

A string giving the value of the filter_variable attribute to filter results by. filter_variable must be set. No filtering by default.

tag_names

A string indicating which series tags to match. Multiple tags can be delimited by a semicolon in a single string (e.g. `"usa;gnp"``).

exclude_tag_names

A string indicating which series tags should not be matched. Multiple tags can be delimited by a semicolon in a single string (e.g. `"usa;gnp"``).

limit

An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset

An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by

A string indicating an attribute by which the results are ordered by. Possible values include:

"series_id" (default)

"title"

"units"

"frequency"

"seasonal_adjustment"

"realtime_start"

"realtime_end"

"last_updated"

"observation_start"

"observation_end"

"popularity"

"group_popularity"

sort_order

A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start

A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end

A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/release/series

See also

Examples

if (fredr_has_key()) {
fredr_release_series(release_id = 20L)

fredr_release_series(release_id = 20L, order_by = "popularity")

# Extract the "catalog" of series from a release on a certain date
fredr_release_series(
   release_id = 20L,
   realtime_end = as.Date("2018-07-13"),
   order_by = "popularity"
)
}
#> # A tibble: 912 × 16
#>    id       realtime_start realtime_end title  observation_start observation_end
#>    <chr>    <chr>          <chr>        <chr>  <chr>             <chr>          
#>  1 ABS1690  2010-02-04     2014-11-05   Asset… 2002-12-18        2014-10-29     
#>  2 ABS1690  2014-11-06     2018-07-13   Asset… 2002-12-18        2014-11-05     
#>  3 D6WABPL  2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#>  4 D1WABPL  2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#>  5 D4WABPL  2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#>  6 D11WABPL 2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#>  7 D9WABPL  2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#>  8 D3WABPL  2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#>  9 D5WABPL  2011-07-07     2018-07-13   Asset… 2002-12-18        2018-07-11     
#> 10 D7WFCDA  2013-04-18     2018-07-13   Asset… 2002-12-18        2018-07-11     
#> # ℹ 902 more rows
#> # ℹ 10 more variables: frequency <chr>, frequency_short <chr>, units <chr>,
#> #   units_short <chr>, seasonal_adjustment <chr>,
#> #   seasonal_adjustment_short <chr>, last_updated <chr>, popularity <int>,
#> #   group_popularity <int>, notes <chr>