Search for a FRED series.

Search for a FRED series.

fredr_search(search_field = "series", ...)

Arguments

search_field
A string indicating to search for a series, series tags, or related series tags. Default is "series" to search for a series. Use "tags" to search tags and "related_tags" to search related tags.

...
A series of paramters to be used in the query. Of the form param_key = 'param_value'. The parameter search_text is required for series/search and series_search_text is required for series/search/tags or series/search/related_tags.

Value

A data.frame object.

Examples

fredr_search(search_text = "oil", order_by = "popularity")

#> # A tibble: 1,000 x 15
#>                   id realtime_start realtime_end
#>                <chr>          <chr>        <chr>
#>  1          CPIAUCSL     2017-05-30   2017-05-30
#>  2        DCOILWTICO     2017-05-30   2017-05-30
#>  3      DCOILBRENTEU     2017-05-30   2017-05-30
#>  4        MCOILWTICO     2017-05-30   2017-05-30
#>  5           WTISPLC     2017-05-30   2017-05-30
#>  6       POILWTIUSDM     2017-05-30   2017-05-30
#>  7       POILBREUSDM     2017-05-30   2017-05-30
#>  8 PCU21311121311101     2017-05-30   2017-05-30
#>  9      MCOILBRENTEU     2017-05-30   2017-05-30
#> 10       IPG211111CN     2017-05-30   2017-05-30
#> # ... with 990 more rows, and 12 more variables: title <chr>,
#> #   observation_start <chr>, observation_end <chr>, frequency <chr>,
#> #   frequency_short <chr>, units <chr>, units_short <chr>,
#> #   seasonal_adjustment <chr>, seasonal_adjustment_short <chr>,
#> #   last_updated <chr>, popularity <int>, notes <chr>