Get all sources of economic data

Get all sources of economic data

fredr_sources(
  ...,
  limit = NULL,
  offset = NULL,
  order_by = NULL,
  sort_order = NULL,
  realtime_start = NULL,
  realtime_end = NULL
)

Arguments

...

These dots only exist for future extensions and should be empty.

limit

An integer limit on the maximum number of results to return. Defaults to 1000, the maximum.

offset

An integer used in conjunction with limit for long series. This mimics the idea of pagination to retrieve large amounts of data over multiple calls. Defaults to 0.

order_by

A string indicating which attribute should be used to order the results. Possible values: "source_id" (default), "name", "realtime_start", "realtime_end".

sort_order

A string representing the order of the resulting series. Possible values are: "asc" (default), and "desc".

realtime_start

A Date indicating the start of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

realtime_end

A Date indicating the end of the real-time period. Defaults to today's date. For more information, see Real-Time Periods.

Value

A tibble object.

API Documentation

fred/sources

See also

Examples

if (fredr_has_key()) {
fredr_sources(limit = 20L)
}
#> # A tibble: 20 × 6
#>       id realtime_start realtime_end name                            link  notes
#>    <int> <chr>          <chr>        <chr>                           <chr> <chr>
#>  1     1 2023-01-23     2023-01-23   Board of Governors of the Fede… http… NA   
#>  2     3 2023-01-23     2023-01-23   Federal Reserve Bank of Philad… http… NA   
#>  3     4 2023-01-23     2023-01-23   Federal Reserve Bank of St. Lo… http… NA   
#>  4     6 2023-01-23     2023-01-23   Federal Financial Institutions… http… NA   
#>  5    11 2023-01-23     2023-01-23   Dow Jones & Company             http… NA   
#>  6    14 2023-01-23     2023-01-23   University of Michigan          http… NA   
#>  7    15 2023-01-23     2023-01-23   Council of Economic Advisers (… http… NA   
#>  8    16 2023-01-23     2023-01-23   U.S. Office of Management and … http… NA   
#>  9    17 2023-01-23     2023-01-23   U.S. Congressional Budget Offi… http… NA   
#> 10    18 2023-01-23     2023-01-23   U.S. Bureau of Economic Analys… http… NA   
#> 11    19 2023-01-23     2023-01-23   U.S. Census Bureau              http… NA   
#> 12    21 2023-01-23     2023-01-23   U.S. Department of Housing and… http… NA   
#> 13    22 2023-01-23     2023-01-23   U.S. Bureau of Labor Statistics http… NA   
#> 14    23 2023-01-23     2023-01-23   U.S. Department of the Treasur… http… NA   
#> 15    26 2023-01-23     2023-01-23   Haver Analytics                 http… NA   
#> 16    31 2023-01-23     2023-01-23   Reserve Bank of Australia       http… NA   
#> 17    32 2023-01-23     2023-01-23   Deutsche Bundesbank             http… NA   
#> 18    33 2023-01-23     2023-01-23   Bank of Italy                   http… Banc…
#> 19    34 2023-01-23     2023-01-23   Swiss National Bank             http… NA   
#> 20    35 2023-01-23     2023-01-23   Central Bank of the Republic o… http… Türk…